ref: update redis, kafka, fix error authentication

+++ src/main/java/com/long1dep/youtuberef11/config/properties/SecurityProperties.java
@@ -15,6 +15,6 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class SecurityProperties {
     String jwtSecret;
-    int jwtExpiration;
-    int rememberMeExpiration;
+    long jwtExpiration;
+    long rememberMeExpiration;
 }
